chore(popup-menu): move entry grouping concern to `PopupMenuComponent`
This ensures we don't double the necessary work.
