+++ lib/features/popup-menu/PopupMenuComponent.js
@@ -14,7 +14,7 @@ import {
 } from 'min-dom';
 
 import PopupMenuHeader from './PopupMenuHeader';
-import PopupMenuList, { groupEntries } from './PopupMenuList';
+import PopupMenuList from './PopupMenuList';
 import classNames from 'clsx';
 import { isDefined, isFunction } from 'min-dash';
 
@@ -22,6 +22,7 @@ import { isDefined, isFunction } from 'min-dash';
  * @typedef {import('./PopupMenuProvider').PopupMenuEntry} PopupMenuEntry
  * @typedef {import('./PopupMenuProvider').PopupMenuHeaderEntry} PopupMenuHeaderEntry
  * @typedef {import('./PopupMenuProvider').PopupMenuEmptyPlaceholderProvider | import('./PopupMenuProvider').PopupMenuEmptyPlaceholder} PopupMenuEmptyPlaceholder
+ * @typedef {import('./PopupMenuProvider').PopupMenuGroup} PopupMenuGroup
  *
  * @typedef {import('../search/search').default} search
  *
@@ -97,6 +98,8 @@ export default function PopupMenuComponent(props) {
   const [ entries, setEntries ] = useState(filterEntries(originalEntries, searchValue));
   const [ selectedEntry, setSelectedEntry ] = useState(entries[0]);
 
+  const groupedEntries = useMemo(() => groupEntries(entries), [ entries ]);
+
   const updateEntries = useCallback((newEntries) => {
 
     // always select first
@@ -112,22 +115,21 @@ export default function PopupMenuComponent(props) {
 
   // handle keyboard selection
   const keyboardSelect = useCallback(direction => {
-    const groupOrderedEntries = orderByGroup(entries);
-    const idx = groupOrderedEntries.indexOf(selectedEntry);
-
+    const entries = getOrderedEntries(groupedEntries);
+    const idx = entries.indexOf(selectedEntry);
 
     let nextIdx = idx + direction;
 
     if (nextIdx < 0) {
-      nextIdx = groupOrderedEntries.length - 1;
+      nextIdx = entries.length - 1;
     }
 
-    if (nextIdx >= groupOrderedEntries.length) {
+    if (nextIdx >= entries.length) {
       nextIdx = 0;
     }
 
-    setSelectedEntry(groupOrderedEntries[nextIdx]);
-  }, [ entries, selectedEntry, setSelectedEntry ]);
+    setSelectedEntry(entries[nextIdx]);
+  }, [ groupedEntries, selectedEntry, setSelectedEntry ]);
 
   const handleKeyDown = useCallback(event => {
     if (event.key === 'Enter' && selectedEntry) {
@@ -197,7 +199,7 @@ export default function PopupMenuComponent(props) {
           ` }
 
           <${PopupMenuList}
-            entries=${ entries }
+            groupedEntries=${ groupedEntries }
             selectedEntry=${ selectedEntry }
             setSelectedEntry=${ setSelectedEntry }
             onAction=${ onSelect }
@@ -297,6 +299,7 @@ function PopupMenuWrapper(props) {
   `;
 }
 
+
 // helpers //////////////////////
 
 function getPopupStyle(props) {
@@ -307,15 +310,45 @@ function getPopupStyle(props) {
   };
 }
 
-function orderByGroup(entries) {
-  const orderedByGroupEntries = [];
-  const groups = groupEntries(entries);
+function getOrderedEntries(groupedEntries) {
+  const entries = [];
 
-  groups.forEach(group => {
+  groupedEntries.forEach(group => {
     group.entries.forEach(entry => {
-      orderedByGroupEntries.push(entry);
+      entries.push(entry);
     });
   });
 
-  return orderedByGroupEntries;
+  return entries;
+}
+
+/**
+ * @param {PopupMenuEntry[]} entries
+ *
+ * @return {PopupMenuGroup[]}
+ */
+export function groupEntries(entries) {
+  const groups = [];
+
+  const getGroup = group => groups.find(elem => group.id === elem.id);
+
+  const containsGroup = group => !!getGroup(group);
+
+  // legacy support for provider built for the old popUp menu
+  const formatGroup = group =>
+    typeof group === 'string' ? { id: group } : group;
+
+  entries.forEach(entry => {
+
+    // assume a default group when none is provided
+    const group = entry.group ? formatGroup(entry.group) : { id: 'default' };
+
+    if (!containsGroup(group)) {
+      groups.push({ ...group, entries: [ entry ] });
+    } else {
+      getGroup(group).entries.push(entry);
+    }
+  });
+
+  return groups;
 }
