+++ lib/features/popup-menu/PopupMenuList.js
@@ -1,6 +1,5 @@
 import {
   html,
-  useMemo,
   useLayoutEffect,
   useRef
 } from '../../ui';
@@ -9,13 +8,14 @@ import PopupMenuItem from './PopupMenuItem';
 
 /**
  * @typedef {import('./PopupMenuProvider').PopupMenuEntry} PopupMenuEntry
+ * @typedef {import('./PopupMenuProvider').PopupMenuGroup} PopupMenuGroup
  */
 
 /**
  * Component that renders a popup menu entry list.
  *
  * @param {Object} props
- * @param {PopupMenuEntry[]} props.entries
+ * @param {PopupMenuGroup[]} props.groupedEntries
  * @param {PopupMenuEntry} props.selectedEntry
  * @param {(entry: PopupMenuEntry | null) => void} props.setSelectedEntry
  */
@@ -23,13 +23,11 @@ export default function PopupMenuList(props) {
   const {
     selectedEntry,
     setSelectedEntry,
-    entries,
+    groupedEntries,
     ...restProps
   } = props;
 
   const resultsRef = useRef();
-
-  const groups = useMemo(() => groupEntries(entries), [ entries ]);
 
   // scroll to selected result
   useLayoutEffect(() => {
@@ -47,7 +45,7 @@ export default function PopupMenuList(props) {
 
   return html`
     <div class="djs-popup-results" ref=${ resultsRef }>
-      ${ groups.map(group => html`
+      ${ groupedEntries.map(group => html`
         ${ group.name && html`
           <div key=${ group.id } class="entry-header" title=${ group.name }>
             ${ group.name }
@@ -71,33 +69,6 @@ export default function PopupMenuList(props) {
 }
 
 
-// helpers
-function groupEntries(entries) {
-  const groups = [];
-
-  const getGroup = group => groups.find(elem => group.id === elem.id);
-
-  const containsGroup = group => !!getGroup(group);
-
-  // legacy support for provider built for the old popUp menu
-  const formatGroup = group =>
-    typeof group === 'string' ? { id: group } : group;
-
-  entries.forEach(entry => {
-
-    // assume a default group when none is provided
-    const group = entry.group ? formatGroup(entry.group) : { id: 'default' };
-
-    if (!containsGroup(group)) {
-      groups.push({ ...group, entries: [ entry ] });
-    } else {
-      getGroup(group).entries.push(entry);
-    }
-  });
-
-  return groups;
-}
-
 // helpers ////////////////
 
 function scrollIntoView(el) {
@@ -110,5 +81,3 @@ function scrollIntoView(el) {
     });
   }
 }
-
-export { groupEntries };
